requirement.txt
aiohttp==3.7.4.post0
akinator.py==4.1.0
async-timeout==3.0.1
asyncpg==0.23.0
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
discord-ext-menus==1.1
discord-py-slash-command==1.1.2
discord.py==1.7.2
discordlists.py==1.1.1
dnspython==2.1.0
h11==0.12.0
httpcore==0.12.3
httpx==0.16.1
humanize==3.7.0
idna==2.10
instant-api-client==0.1.2
mpmath==1.2.1
multidict==5.1.0
numpy==1.20.3
olefile==0.46
opencv-python==4.5.2.52
Pillow==4.2.1
prsaw==0.3.1
psutil==5.8.0
pymongo==3.11.4
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
rfc3986==1.5.0
sniffio==1.2.0
sr-api==1.1.3
statcord.py==3.1.0
termcolor==1.1.0
typing-extensions==3.10.0.0
urllib3==1.26.5
Wand==0.6.6
yarl==1.6.3
youtube-dl==2021.5.16
